fix: تغییر DEF_AMOUNT در PAY2_ITEM_TMPL_LINE به DECIMAL(18,2)
مشکل: قالب‌های حکم (PAY2_ITEM_TMPL_LINE) هنگام ایجاد حکم جدید، مقدار
DEF_AMOUNT را مستقیماً به AMOUNT در PAY2_DECREE_LINE کپی می‌کردند.
اگر DEF_AMOUNT همچنان BIGINT می‌ماند، مقادیر اعشاری حق شیفت (مثل 7.5)
هنگام کپی گرد می‌شدند و اطلاعات از دست می‌رفت.

تغییرات:
- Migration 008: افزودن idempotent migration برای PAY2_ITEM_TMPL_LINE.DEF_AMOUNT
- ScriptSqly.cs: DEF_AMOUNT DECIMAL(18,2) برای نصب تازه
- PAY2_DDL_v6.sql: به‌روزرسانی فایل مرجع
- Pay2ItemTmplLineDto.DEF_AMOUNT: long → decimal
- ItemTemplateLineModal: decimal.TryParse در SubmitSave
- ItemTemplateLineModal: اصلاح AmountStr در EditLine (جلوگیری از باگ ×100)

+++ Server/Info/ScriptSqly.cs
@@ -596,7 +596,7 @@ CREATE TABLE [dbo].[PAY2_ITEM_TMPL_LINE]
 (
     [TMPL_ID]   INT      NOT NULL,
     [ITEM_ID]   INT      NOT NULL,
-    [DEF_AMOUNT] BIGINT  NOT NULL CONSTRAINT DF_TL_AMT DEFAULT(0),
+    [DEF_AMOUNT] DECIMAL(18,2) NOT NULL CONSTRAINT DF_TL_AMT DEFAULT(0),
     [INS_OV]    BIT      NULL,                                               -- NULL=از تعریف آیتم
     [TAX_OV]    BIT      NULL,
     [BASIS_OV]  TINYINT  NULL,
